modify fragment near view

+++ app/src/main/java/com/hwl/beta/net/near/body/AddNearCommentRequest.java
@@ -1,46 +1,42 @@
 package com.hwl.beta.net.near.body;
 
-/**
- * Created by Administrator on 2018/3/14.
- */
-
 public class AddNearCommentRequest {
 
     /// <summary>
     /// 附近信息id
     /// </summary>
-    private int NearCircleId;
+    private long NearCircleId;
     /// <summary>
     /// 发评论的用户id
     /// </summary>
-    private int CommentUserId;
+    private long CommentUserId;
     /// <summary>
     /// 回复用户id
     /// </summary>
-    private int ReplyUserId;
+    private long ReplyUserId;
     private String Content;
 
-    public int getNearCircleId() {
+    public long getNearCircleId() {
         return NearCircleId;
     }
 
-    public void setNearCircleId(int nearCircleId) {
+    public void setNearCircleId(long nearCircleId) {
         NearCircleId = nearCircleId;
     }
 
-    public int getCommentUserId() {
+    public long getCommentUserId() {
         return CommentUserId;
     }
 
-    public void setCommentUserId(int commentUserId) {
+    public void setCommentUserId(long commentUserId) {
         CommentUserId = commentUserId;
     }
 
-    public int getReplyUserId() {
+    public long getReplyUserId() {
         return ReplyUserId;
     }
 
-    public void setReplyUserId(int replyUserId) {
+    public void setReplyUserId(long replyUserId) {
         ReplyUserId = replyUserId;
     }
